Disable some tests until solve concurrency issue.
